Update getPageTerms.php
Have added an optional placeholder (for output) and an optional 'count' placeholder (for innerTpl) in this version. Using this in a site currently under development with no problem. Has enabled us to generate output (headings and surrounding panels etc) only if there are terms associated with a particular resource page. Also (using a conditional output filter in the innerTpl) the ability to create a hyperlink to 'term' page (listing associated resources) only if there is more than one resource document associated. Hope that makes sense. You may be able to improve on the SQL statement when 'count' enabled. Not my finest hour :)